feat: 1:1문의 레이아웃 구현
