+++ src/components/common/Banner.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ListBanner } from '../../../styles/common/List-styled';
+import { ListBanner } from '../../styles/common/List-styled';
 
 const Banner = ({data, text}) => {
   return <ListBanner backgroundImage = {data}>{text}</ListBanner>;
